feat: remoção de pedido de acordo com o estado dele

+++ src/main/java/com/ufcg/psoft/pitsA/service/pedido/PedidoRemoverServiceImpl.java
@@ -1,5 +1,8 @@
 package com.ufcg.psoft.pitsA.service.pedido;
 
+import com.ufcg.psoft.pitsA.exception.pedido.PedidoJaEstaProntoException;
+import com.ufcg.psoft.pitsA.exception.pedido.PedidoNaoEncontradoException;
+import com.ufcg.psoft.pitsA.model.pedido.Pedido;
 import com.ufcg.psoft.pitsA.repository.PedidoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -11,6 +14,10 @@ public class PedidoRemoverServiceImpl implements PedidoRemoverService {
 
     @Override
     public void removerPedido(Long pedidoId) {
+        Pedido pedido = pedidoRepository.findById(pedidoId).orElseThrow(PedidoNaoEncontradoException::new);
+        if (pedido.getEstado().isPronto() || pedido.getEstado().isRota() || pedido.getEstado().isEntregue())
+            throw new PedidoJaEstaProntoException();
+
         pedidoRepository.deleteById(pedidoId);
     }
 }
